+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-## Changelog
+# Changelog
+All notable changes to this extension will be documented in this file.
 
-### 0.5.1 (2018-11-21)
-* [BUGFXI] fixed translation for file uploader
+## [0.5.2] - 2018-12-14
+### Added
+- added Site configuration duplication
+- added google font url validator
+
+## [0.5.1] - 2018-11-21
+### Added
+- translation for file uploader